% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sembag.R
\name{sembag_inloop}
\alias{sembag_inloop}
\title{Title}
\usage{
sembag_inloop(
  iteration = 1,
  data,
  formula,
  iterations,
  fit_function = NULL,
  variable_sampler = NULL,
  data_sampler = NULL,
  validation_function = NULL,
  mtry = NULL,
  ...
)
}
\arguments{
\item{data}{the dataset}

\item{formula}{A formula of the form ~ a + b + c. Note, there is no need to
put an outcome variable in the formula}

\item{iterations}{number of iterations}

\item{fit_function}{a custom function to use for estimating parameters}
}
\value{
a list
}
\description{
Title
}
